+++ src/spdx_tools/spdx3/parser/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2026-present SPDX contributors
+# SPDX-License-Identifier: Apache-2.0
+from spdx_tools.spdx3.parser.parse_anything import parse_file
+
+__all__ = ["parse_file"]